% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_negllh_allGSI.R
\name{flex_negllh_allGSI}
\alias{flex_negllh_allGSI}
\title{calculate the negative log-likelihood with only PBT and GSI (optional) information}
\usage{
flex_negllh_allGSI(params, nPBT, nGSI, ohnc, t, utGSI, ohnc_gsi, pbtGSIkey)
}
\arguments{
\item{params}{list of paramaters to optimize}

\item{nPBT}{number of PBT groups to estimate}

\item{nGSI}{number of GSI groups to estimate}

\item{ohnc}{list of number of observed (PBT assigned) hatchery fish in each PBT and GSI group (gsi groups should be 0)}

\item{t}{list of tag rates for all PBT and GSI groups (gsi groups should be 0)}

\item{utGSI}{list of number of un-PBT assigned fish in each GSI group}

\item{ohnc_gsi}{matrix of counts of fish GSI assigned to various groups}

\item{pbtGSIkey}{list with a vector telling which GSI group (as integer giving their position in the order)
is fixed at 1 for each pbt group (shoudl be a non-zero ohnc_gsi group)}
}
\description{
this function does not allow another variable. for that, see \code{flex_negllh_var}
}
