% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatePBTGSIdata.R
\name{generatePBTGSIdata}
\alias{generatePBTGSIdata}
\title{This function generates data for one strata.
To get multiple strata, run repeatedly and add strata variable}
\usage{
generatePBTGSIdata(sampRate = 0.1, censusSize = 3000,
  relSizePBTgroups = 1, tagRates = 0.8, obsTagRates = 0.8,
  physTagRates = 0, true_clipped = 0.5, true_noclip_H = 0.2,
  true_wild = 0.3, relSizeGSIgroups = 1, PBT_GSI_calls = 1,
  varMatList = NA)
}
\arguments{
\item{PBT_GSI_calls}{matrix with rows equal to PBT groups, columns equal to GSI group and
values giving the proportions of teh PBT fish that GSI assign to that GSI group}

\item{varMatList}{list of matrices, one for each variable. rows are PBT/GSI groups, in order of PBT groups followed
by GSI groups, columns being the categories for the variable, and values beign the proportion of the
group that has that variable value}
}
\description{
This function generates data for one strata.
To get multiple strata, run repeatedly and add strata variable
}
