% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepOneStrata.R
\name{prepOneStrata}
\alias{prepOneStrata}
\title{This function takes data in a standard table format and prepares it for one strata for MCpbt}
\usage{
prepOneStrata(trapData, tags, GSIcol, PBTcol, variableCols = c(),
  variableColsOth = c(), adFinCol, AI = TRUE, verbose = TRUE,
  GSIgroups = NA, variableValues = NA, variableValuesOth = NA,
  strataName = NA)
}
\arguments{
\item{trapData}{dataframe with data for fish sampled from one strata - trap data for dam escapement}

\item{tags}{dataframe with tag rates for all groups. This should NOT include the Unassigned group.}

\item{GSIcol}{column name of column containing GSI assignments. If NA, the function gives all fish the same gsi assignment,
which is equivalent to not having GSI information.}

\item{PBTcol}{column name of column containing PBT assignments}

\item{variableCols}{vector of column names of columns containing the variables to estimate composition and 
to inform the group selection for PBT untagged fish}

\item{variableColsOth}{vector of column names of columns containing the variables to estimate composition but 
NOT inform the group selection for PBT untagged fish}

\item{adFinCol}{column name of column containing adipose fin status - TRUE being intact FALSE being clipped, NA missing}

\item{AI}{TRUE to analyse ad-intact fish, FALSE to analyze ad-clipped fish}

\item{verbose}{TRUE to print some messages, FALSE to not}

\item{GSIgroups}{These are the values for all the GSI groups that you expect to be present in the population. If NA, then
the values are taken to be all the unique values in the GSI column in the dataset.}

\item{variableValues}{a list, in the order of \code{variableCols} with entries having the values expected for each variable. This
is helpful to make sure the same variable values are estiamted in each strata even if one value is not observed in all strata. If
a value other than a list is given (for example, NA), it uses the values present in the dataset.}

\item{variableValuesOth}{Same as variableValues, but for variableColsOth}
}
\description{
This function takes data in a standard table format and prepares it for one strata for MCpbt
}
