% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepStrata.R
\name{prepStrata}
\alias{prepStrata}
\title{This function prepares the data to run multiple strata}
\usage{
prepStrata(trapData, tags, GSIcol, PBTcol, strataCol, variableCols = c(),
  variableColsOth = c(), adFinCol, AI = TRUE, GSIgroups = NA,
  variableValues = NA, variableValuesOth = NA, verbose = TRUE)
}
\arguments{
\item{trapData}{dataframe with data for fish sampled from one strata - trap data for dam escapement}

\item{tags}{dataframe with tag rates for all groups. This should NOT include the Unassigned group.}

\item{GSIcol}{column name of column containing GSI assignments. If NA, the function gives all fish the same gsi assignment,
which is equivalent to not having GSI information.}

\item{PBTcol}{column name of column containing PBT assignments}

\item{strataCol}{column name of column indicating the strata the observation belongs to}

\item{variableCols}{vector of column names of columns containing the variables to estimate composition and 
to inform the group selection for PBT untagged fish}

\item{variableColsOth}{vector of column names of columns containing the variables to estimate composition but 
NOT inform the group selection for PBT untagged fish}

\item{adFinCol}{column name of column containing adipose fin status - TRUE being intact FALSE being clipped, NA missing}

\item{AI}{TRUE to analyze ad-intact fish, FALSE to analyze ad-clipped fish}

\item{GSIgroups}{These are the values for all the GSI groups that you expect to be present in the population. If NA, then
the values are taken to be all the unique values in the GSI column in the dataset within each strata (ie unobserved groups are
assumed to be 0).}

\item{variableValues}{a list, in the order of \code{variableCols} with entries having the values expected for each variable. This
is helpful to make sure the same variable values are estiamted in each strata even if one value is not observed in all strata. If
a value other than a list is given (for example, NA), it uses the values present in the dataset within each strata (ie unobserved 
groups are assumed to be 0).}

\item{variableValuesOth}{Same as variableValues, but for variableColsOth}

\item{verbose}{TRUE to print some messages, FALSE to not}
}
\description{
The model is actually run by a separate function in order to allow
a user to edit the priors and initial values before running.
}
