% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearTypeData.R
\name{rearTypeData}
\alias{rearTypeData}
\title{H HNC W data simulation function}
\usage{
rearTypeData(sampRates = 0.1, censusSizes = 3000,
  relSizePBTgroups = 1, tagRates = 0.8, obsTagRates = 0.8,
  physTagRates = 0, true_clipped = 0.5, true_noclip_H = 0.2,
  true_wild = 0.3)
}
\description{
H HNC W data simulation function
}
