% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearTypeMultVarData.R
\name{rearTypeMultVarData}
\alias{rearTypeMultVarData}
\title{H HNC W data simulation function}
\usage{
rearTypeMultVarData(sampRates = 0.1, censusSizes = 3000,
  relSizePBTgroups = 1, tagRates = 0.8, obsTagRates = 0.8,
  physTagRates = 0, true_clipped = 0.5, true_noclip_H = 0.2,
  true_wild = 0.3, varArray = NA)
}
\arguments{
\item{varArray}{list of arrays with first dimension for each PBT groups (in the same order
as relSizePBTgroups) plus wild as the last entry of the first dimension
second dimension for each strata, and third dimension for each category in the variable.
Values are the proportion of the strata and PBT group that belong to each category of the variable}
}
\description{
H HNC W data simulation function
}
